[svlogd] prevent deletion of the processing file
svlogd expects time to be correct, but sometimes this is not the case. If time
information is inaccurate, it is possible for svlogd to delete the file it is
about to process. When this happens, svlogd will enter an infinite busy loop
trying to process a file that no longer exists. This patch prevents that case
from happening by excluding processing file from consideration when removing old
log files.